Refactor EventCreator packaging, and rename classes as suggested in #158
