+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             </intent-filter>
 
         </activity> <!-- Why on god's green earth are those lines needed for EventCreatorTest to pass ? -->
-        <activity android:name=".events.EventCreator" />
+        <activity android:name=".events.ui.EventCreator" />
         <activity android:name=".EventsView" />
 
         <receiver android:name=".notifs.NotificationPublisher" />
