+++ app/src/main/java/com/github/onedirection/EventsView.java
@@ -7,7 +7,7 @@ import android.os.Bundle;
 import android.widget.TextView;
 
 import com.github.onedirection.events.Event;
-import com.github.onedirection.events.EventCreator;
+import com.github.onedirection.events.ui.EventCreator;
 import com.github.onedirection.notifs.Notifications;
 
 import java.util.HashSet;
