+++ app/src/main/res/layout/event_creator.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".events.EventCreator">
+    tools:context=".events.ui.EventCreator">
 
     <ScrollView
         android:layout_width="match_parent"
