+++ app/src/main/res/layout/event_creator_geolocation_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".events.EventCreatorGeolocationFragment">
+    tools:context=".events.ui.GeolocationFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
