+++ app/src/main/res/layout/event_creator_main_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".events.EventCreatorMainFragment">
+    tools:context=".events.ui.MainFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
